Update
-Ejercicio 1 done
-Ejercicio 2 done
-Ejercicio 3 done
-Ejercicio 4 iniciado

+++ Ejercicio1/src/main/java/com/crud/api/dao/IArticulosDAO.java
@@ -1,8 +1,9 @@
 package com.crud.api.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.crud.api.dto.Articulos;
 
-public interface IArticulosDAO extends JpaRepository<Articulos, Long>{
+public interface IArticulosDAO extends JpaRepository<Articulos, Integer> {
 
 }
